Move Glossary model to own app. Rename to GlossaryItem
