+++ backend/api/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from api.models import Category, Glossary, Question, Quiz, QuizQuestion, Tag
+from api.models import Category, Question, Quiz, QuizQuestion, Tag
 
 
 class SimpleChoiceSerializer(serializers.Serializer):
@@ -178,22 +178,3 @@ class QuizFullSerializer(serializers.ModelSerializer):
     class Meta:
         model = Quiz
         fields = QUIZ_FIELDS
-
-
-"""
-GLOSSARY
-"""
-
-
-class GlossarySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Glossary
-        fields = [
-            "name",
-            "name_alternatives",
-            "definition_short",
-            "description",
-            "description_accessible_url",
-            "added",
-            "created",
-        ]
